Early work on deformation driven correspondence.

+++ ddtt/StructureGraphLib/Dijkstra.h
@@ -29,7 +29,7 @@ static void DijkstraComputePaths(vertex_t source,
 	std::vector<weight_t> &min_distance,
 	std::vector<vertex_t> &previous)
 {
-	int n = adjacency_list.size();
+	size_t n = adjacency_list.size();
 	min_distance.clear();
 	min_distance.resize(n, max_weight);
 	min_distance[source] = 0;
